refactor: App 클래스에서 GameRunner와 GameInitializer로 분리
